Restructure classes
