+++ app/src/main/res/navigation/main_graph.xml
@@ -455,7 +455,7 @@
 
     <fragment
         android:id="@+id/nftsFragment"
-        android:name="io.horizontalsystems.bankwallet.modules.nft.NftsFragment" />
+        android:name="io.horizontalsystems.bankwallet.modules.nft.collection.NftCollectionsFragment" />
 
     <!--  Dialogs/BottomSheets  -->
     <!--    <dialog-->
